requirements.txt
feedparser
requests
beautifulsoup4
openpyxl
pyyaml
python-dotenv
flask
